feat : permit the order of hero-sections
