+++ app/preview/hero-section/two/page.tsx
@@ -1,114 +1,226 @@
-'use client'
-import { Logo } from '@/components/logo'
-import { Button } from '@/components/ui/button'
-import { ArrowRight, Mail, Menu, SendHorizonal, X } from 'lucide-react'
-import Image from 'next/image'
+import React from 'react'
 import Link from 'next/link'
-import { useState } from 'react'
+import { ChevronRight } from 'lucide-react'
+import { Button } from '@/components/ui/button'
+import Image from 'next/image'
+import { TextEffect } from '@/components/motion-primitives/text-effect'
+import { AnimatedGroup } from '@/components/motion-primitives/animated-group'
+import { HeroHeader } from '@/components/hero6-header'
 
-const menuItems = [
-    { name: 'Features', href: '#' },
-    { name: 'Solution', href: '#' },
-    { name: 'Pricing', href: '#' },
-    { name: 'About', href: '#' },
-]
+const transitionVariants = {
+    item: {
+        hidden: {
+            opacity: 0,
+            filter: 'blur(12px)',
+            y: 12,
+        },
+        visible: {
+            opacity: 1,
+            filter: 'blur(0px)',
+            y: 0,
+            transition: {
+                type: 'spring',
+                bounce: 0.3,
+                duration: 1.5,
+            },
+        },
+    },
+}
 
 export default function HeroSection() {
-    const [menuState, setMenuState] = useState(false)
     return (
         <>
-            <header>
-                <nav data-state={menuState && 'active'} className="fixed z-20 w-full border-b border-dashed bg-white backdrop-blur md:relative dark:bg-zinc-950/50 lg:dark:bg-transparent">
-                    <div className="m-auto max-w-5xl px-6">
-                        <div className="flex flex-wrap items-center justify-between gap-6 py-3 lg:gap-0 lg:py-4">
-                            <div className="flex w-full justify-between lg:w-auto">
-                                <Link href="/" aria-label="home" className="flex items-center space-x-2">
-                                    <Logo />
-                                </Link>
+            <HeroHeader />
+            <main className="overflow-hidden">
+                <div
+                    aria-hidden
+                    className="absolute inset-0 isolate hidden contain-strict lg:block">
+                    <div className="w-140 h-320 -translate-y-87.5 absolute left-0 top-0 -rotate-45 rounded-full bg-[radial-gradient(68.54%_68.72%_at_55.02%_31.46%,hsla(0,0%,85%,.08)_0,hsla(0,0%,55%,.02)_50%,hsla(0,0%,45%,0)_80%)]" />
+                    <div className="h-320 absolute left-0 top-0 w-60 -rotate-45 rounded-full bg-[radial-gradient(50%_50%_at_50%_50%,hsla(0,0%,85%,.06)_0,hsla(0,0%,45%,.02)_80%,transparent_100%)] [translate:5%_-50%]" />
+                    <div className="h-320 -translate-y-87.5 absolute left-0 top-0 w-60 -rotate-45 bg-[radial-gradient(50%_50%_at_50%_50%,hsla(0,0%,85%,.04)_0,hsla(0,0%,45%,.02)_80%,transparent_100%)]" />
+                </div>
+                <section>
+                    <div className="relative pt-24">
+                        <div className="absolute inset-0 -z-10 size-full [background:radial-gradient(125%_125%_at_50%_100%,transparent_0%,var(--color-background)_75%)]"></div>
+                        <div className="mx-auto max-w-5xl px-6">
+                            <div className="sm:mx-auto lg:mr-auto lg:mt-0">
+                                <TextEffect
+                                    preset="fade-in-blur"
+                                    speedSegment={0.3}
+                                    as="h1"
+                                    className="mt-8 max-w-2xl text-balance text-5xl font-medium md:text-6xl lg:mt-16">
+                                    Build and Ship 10x faster with NS
+                                </TextEffect>
+                                <TextEffect
+                                    per="line"
+                                    preset="fade-in-blur"
+                                    speedSegment={0.3}
+                                    delay={0.5}
+                                    as="p"
+                                    className="mt-8 max-w-2xl text-pretty text-lg">
+                                    Tailwindcss highly customizable components for building modern websites and applications that look and feel the way you mean it.
+                                </TextEffect>
 
-                                <button onClick={() => setMenuState(!menuState)} aria-label={menuState == true ? 'Close Menu' : 'Open Menu'} className="relative z-20 -m-2.5 -mr-4 block cursor-pointer p-2.5 lg:hidden">
-                                    <Menu className="in-data-[state=active]:rotate-180 in-data-[state=active]:scale-0 in-data-[state=active]:opacity-0 m-auto size-6 duration-200" />
-                                    <X className="in-data-[state=active]:rotate-0 in-data-[state=active]:scale-100 in-data-[state=active]:opacity-100 absolute inset-0 m-auto size-6 -rotate-180 scale-0 opacity-0 duration-200" />
-                                </button>
-                            </div>
-
-                            <div className="bg-background in-data-[state=active]:block lg:in-data-[state=active]:flex mb-6 hidden w-full flex-wrap items-center justify-end space-y-8 rounded-3xl border p-6 shadow-2xl shadow-zinc-300/20 md:flex-nowrap lg:m-0 lg:flex lg:w-fit lg:gap-6 lg:space-y-0 lg:border-transparent lg:bg-transparent lg:p-0 lg:shadow-none dark:shadow-none dark:lg:bg-transparent">
-                                <div className="lg:pr-4">
-                                    <ul className="space-y-6 text-base lg:flex lg:gap-8 lg:space-y-0 lg:text-sm">
-                                        {menuItems.map((item, index) => (
-                                            <li key={index}>
-                                                <Link href={item.href} className="text-muted-foreground hover:text-accent-foreground block duration-150">
-                                                    <span>{item.name}</span>
-                                                </Link>
-                                            </li>
-                                        ))}
-                                    </ul>
-                                </div>
-
-                                <div className="flex w-full flex-col space-y-3 sm:flex-row sm:gap-3 sm:space-y-0 md:w-fit lg:border-l lg:pl-6">
-                                    <Button asChild variant="outline" size="sm">
-                                        <Link href="#">
-                                            <span>Login</span>
+                                <AnimatedGroup
+                                    variants={{
+                                        container: {
+                                            visible: {
+                                                transition: {
+                                                    staggerChildren: 0.05,
+                                                    delayChildren: 0.75,
+                                                },
+                                            },
+                                        },
+                                        ...transitionVariants,
+                                    }}
+                                    className="mt-12 flex items-center gap-2">
+                                    <div
+                                        key={1}
+                                        className="bg-foreground/10 rounded-[calc(var(--radius-xl)+0.125rem)] border p-0.5">
+                                        <Button
+                                            asChild
+                                            size="lg"
+                                            className="rounded-xl px-5 text-base">
+                                            <Link href="#link">
+                                                <span className="text-nowrap">Start Building</span>
+                                            </Link>
+                                        </Button>
+                                    </div>
+                                    <Button
+                                        key={2}
+                                        asChild
+                                        size="lg"
+                                        variant="ghost"
+                                        className="h-10.5 rounded-xl px-5 text-base">
+                                        <Link href="#link">
+                                            <span className="text-nowrap">Request a demo</span>
                                         </Link>
                                     </Button>
-
-                                    <Button asChild size="sm">
-                                        <Link href="#">
-                                            <span>Login</span>
-                                        </Link>
-                                    </Button>
-                                </div>
+                                </AnimatedGroup>
                             </div>
                         </div>
+                        <AnimatedGroup
+                            variants={{
+                                container: {
+                                    visible: {
+                                        transition: {
+                                            staggerChildren: 0.05,
+                                            delayChildren: 0.75,
+                                        },
+                                    },
+                                },
+                                ...transitionVariants,
+                            }}>
+                            <div className="relative -mr-56 mt-8 overflow-hidden px-2 sm:mr-0 sm:mt-12 md:mt-20">
+                                <div
+                                    aria-hidden
+                                    className="bg-linear-to-b to-background absolute inset-0 z-10 from-transparent from-35%"
+                                />
+                                <div className="inset-shadow-2xs ring-background dark:inset-shadow-white/20 bg-background relative mx-auto max-w-5xl overflow-hidden rounded-2xl border p-4 shadow-lg shadow-zinc-950/15 ring-1">
+                                    <Image
+                                        className="bg-background aspect-15/8 relative hidden rounded-2xl dark:block"
+                                        src="/mail2.png"
+                                        alt="app screen"
+                                        width="2700"
+                                        height="1440"
+                                    />
+                                    <Image
+                                        className="z-2 border-border/25 aspect-15/8 relative rounded-2xl border dark:hidden"
+                                        src="/mail2-light.png"
+                                        alt="app screen"
+                                        width="2700"
+                                        height="1440"
+                                    />
+                                </div>
+                            </div>
+                        </AnimatedGroup>
                     </div>
-                </nav>
-            </header>
+                </section>
+                <section className="bg-background pb-16 pt-16 md:pb-32">
+                    <div className="group relative m-auto max-w-5xl px-6">
+                        <div className="absolute inset-0 z-10 flex scale-95 items-center justify-center opacity-0 duration-500 group-hover:scale-100 group-hover:opacity-100">
+                            <Link
+                                href="/"
+                                className="block text-sm duration-150 hover:opacity-75">
+                                <span> Meet Our Customers</span>
 
-            <main>
-                <section className="overflow-hidden">
-                    <div className="relative mx-auto max-w-5xl px-6 py-28 lg:py-20">
-                        <div className="lg:flex lg:items-center lg:gap-12">
-                            <div className="relative z-10 mx-auto max-w-xl text-center lg:ml-0 lg:w-1/2 lg:text-left">
-                                <Link href="/" className="rounded-(--radius) mx-auto flex w-fit items-center gap-2 border p-1 pr-3 lg:ml-0">
-                                    <span className="bg-muted rounded-[calc(var(--radius)-0.25rem)] px-2 py-1 text-xs">New</span>
-                                    <span className="text-sm">Introduction Tailus UI Html</span>
-                                    <span className="bg-(--color-border) block h-4 w-px"></span>
-
-                                    <ArrowRight className="size-4" />
-                                </Link>
-
-                                <h1 className="mt-10 text-balance text-4xl font-bold md:text-5xl xl:text-5xl">Production Ready Digital Marketing blocks</h1>
-                                <p className="mt-8">Error totam sit illum. Voluptas doloribus asperiores quaerat aperiam. Quidem harum omnis beatae ipsum soluta!</p>
-
-                                <div>
-                                    <form action="" className="mx-auto my-10 max-w-sm lg:my-12 lg:ml-0 lg:mr-auto">
-                                        <div className="bg-background has-[input:focus]:ring-muted relative grid grid-cols-[1fr_auto] items-center rounded-[calc(var(--radius)+0.75rem)] border pr-3 shadow shadow-zinc-950/5 has-[input:focus]:ring-2">
-                                            <Mail className="text-caption pointer-events-none absolute inset-y-0 left-5 my-auto size-5" />
-
-                                            <input placeholder="Your mail address" className="h-14 w-full bg-transparent pl-12 focus:outline-none" type="email" />
-
-                                            <div className="md:pr-1.5 lg:pr-0">
-                                                <Button aria-label="submit" className="rounded-(--radius)">
-                                                    <span className="hidden md:block">Get Started</span>
-                                                    <SendHorizonal className="relative mx-auto size-5 md:hidden" strokeWidth={2} />
-                                                </Button>
-                                            </div>
-                                        </div>
-                                    </form>
-
-                                    <ul className="list-inside list-disc space-y-2">
-                                        <li>Faster</li>
-                                        <li>Modern</li>
-                                        <li>100% Customizable</li>
-                                    </ul>
-                                </div>
-                            </div>
+                                <ChevronRight className="ml-1 inline-block size-3" />
+                            </Link>
                         </div>
-                        <div className="absolute inset-0 -mx-4 rounded-3xl p-3 lg:col-span-3">
-                            <div className="relative">
-                                <div className="bg-radial-[at_65%_25%] to-background z-1 -inset-17 absolute from-transparent to-40%"></div>
-                                <Image className="hidden dark:block" src="/music.png" alt="app illustration" width={2796} height={2008} />
-                                <Image className="dark:hidden" src="/music-light.png" alt="app illustration" width={2796} height={2008} />
+                        <div className="group-hover:blur-xs mx-auto mt-12 grid max-w-2xl grid-cols-4 gap-x-12 gap-y-8 transition-all duration-500 group-hover:opacity-50 sm:gap-x-16 sm:gap-y-14">
+                            <div className="flex">
+                                <img
+                                    className="mx-auto h-5 w-fit dark:invert"
+                                    src="https://html.tailus.io/blocks/customers/nvidia.svg"
+                                    alt="Nvidia Logo"
+                                    height="20"
+                                    width="auto"
+                                />
+                            </div>
+
+                            <div className="flex">
+                                <img
+                                    className="mx-auto h-4 w-fit dark:invert"
+                                    src="https://html.tailus.io/blocks/customers/column.svg"
+                                    alt="Column Logo"
+                                    height="16"
+                                    width="auto"
+                                />
+                            </div>
+                            <div className="flex">
+                                <img
+                                    className="mx-auto h-4 w-fit dark:invert"
+                                    src="https://html.tailus.io/blocks/customers/github.svg"
+                                    alt="GitHub Logo"
+                                    height="16"
+                                    width="auto"
+                                />
+                            </div>
+                            <div className="flex">
+                                <img
+                                    className="mx-auto h-5 w-fit dark:invert"
+                                    src="https://html.tailus.io/blocks/customers/nike.svg"
+                                    alt="Nike Logo"
+                                    height="20"
+                                    width="auto"
+                                />
+                            </div>
+                            <div className="flex">
+                                <img
+                                    className="mx-auto h-5 w-fit dark:invert"
+                                    src="https://html.tailus.io/blocks/customers/lemonsqueezy.svg"
+                                    alt="Lemon Squeezy Logo"
+                                    height="20"
+                                    width="auto"
+                                />
+                            </div>
+                            <div className="flex">
+                                <img
+                                    className="mx-auto h-4 w-fit dark:invert"
+                                    src="https://html.tailus.io/blocks/customers/laravel.svg"
+                                    alt="Laravel Logo"
+                                    height="16"
+                                    width="auto"
+                                />
+                            </div>
+                            <div className="flex">
+                                <img
+                                    className="mx-auto h-7 w-fit dark:invert"
+                                    src="https://html.tailus.io/blocks/customers/lilly.svg"
+                                    alt="Lilly Logo"
+                                    height="28"
+                                    width="auto"
+                                />
+                            </div>
+
+                            <div className="flex">
+                                <img
+                                    className="mx-auto h-6 w-fit dark:invert"
+                                    src="https://html.tailus.io/blocks/customers/openai.svg"
+                                    alt="OpenAI Logo"
+                                    height="24"
+                                    width="auto"
+                                />
                             </div>
                         </div>
                     </div>
